refactor: move check-limit methods to their own module
