+++ src/app/payments/index.ts
@@ -1,6 +1,7 @@
-export * from "./get-protocol-fee"
-export * from "./send-lightning"
-export * from "./send-intraledger"
-export * from "./update-pending-payments"
-export * from "./reimburse-fee"
 export * from "./add-earn"
+export * from "./get-protocol-fee"
+export * from "./limits-check"
+export * from "./reimburse-fee"
+export * from "./send-intraledger"
+export * from "./send-lightning"
+export * from "./update-pending-payments"
